requirements.txt
# Core libraries
numpy==1.24.3
pandas==2.0.3
scipy==1.11.1

# Deep learning frameworks
torch==2.0.1
torchvision==0.15.2
tensorflow==2.13.0
transformers==4.31.0

# Image processing
pillow==10.0.0
opencv-python-headless==4.8.0.74
albumentations==1.3.1

# Data visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Machine learning
scikit-learn==1.3.0
faiss-cpu==1.7.4  # Use faiss-gpu for CUDA support

# Utilities
tqdm==4.65.0
h5py==3.9.0
jupyter==1.0.0
ipywidgets==8.0.6  # For interactive widgets in notebooks
